+++ app/src/main/java/co/develhope/meteoapp/adapter/RecentSearchAdapter.kt
@@ -1,16 +1,12 @@
 package co.develhope.meteoapp.adapter
 
-import android.text.Layout
+
 import android.view.LayoutInflater
-import android.view.View
 import android.view.ViewGroup
-import android.widget.TextView
 import androidx.recyclerview.widget.RecyclerView
-import androidx.recyclerview.widget.RecyclerView.Recycler
 import co.develhope.meteoapp.R
-import co.develhope.meteoapp.data.RecentSearchDataSet
-
-
+import co.develhope.meteoapp.adapter.holder.RecentSearchTitleHolder
+import co.develhope.meteoapp.adapter.holder.SearchTermViewHolder
 import co.develhope.meteoapp.model.RecyclerViewData
 
 const val SEARCHITEM = 1
@@ -20,21 +16,9 @@ const val SEARCHTITLE = 0
 class RecentSearchAdapter(val data: List<RecyclerViewData>) :
     RecyclerView.Adapter<RecyclerView.ViewHolder>() {
 
-    class RecentSearchTitleHolder(view : View) : RecyclerView.ViewHolder(view){
-        private val title = view.findViewById<TextView>(R.id.recent_search_title)
 
-        fun bind(searchtitle : RecyclerViewData.RecentSearchTitle){
-           title.text = searchtitle.title
-        }
-    }
 
-    class SearchTermViewHolder(view: View) : RecyclerView.ViewHolder(view) {
-        private val searchTerm = view.findViewById<TextView>(R.id.citynametextview)
 
-        fun bind(city : RecyclerViewData.RecentSearch){
-            searchTerm.text = city.cityName
-        }
-    }
 
     override fun getItemViewType(position: Int): Int {
         return when(data[position]){
